[#3794] Fold email/ID nets into scrub_url; consolidate scrubbing specs
Issue #3794 C2/C4: scrub_url now runs the email and verifiable-identifier
nets after the path and named-param passes, making it the single thorough
tier for every URL-shaped value (request.url, Referer, transaction name,
span url/http.query); scrub_text becomes a semantic alias.
scrub_query_string tolerates a leading '?' (same bug as frontend C1) and
preserves it. sentry_url_scrubbing_spec.rb is merged into
setup_diagnostics_spec.rb so the scrubbing coverage lives with the
initializer it tests.
